fixed liek 5+ bugs

+++ Assets/Project/Scripts/Card Data Scripts/ExtractionCardData.cs
@@ -7,7 +7,7 @@ public class ExtractionCardData : CardData
     [HorizontalLine("Base Card Data Settings", 3, FixedColor.DustyBlue)]
     public RawResource ResourceToExtract;
     [Preview][SerializeField] public GameObject resourcePrefab;  // Prefab for the resource to spawn
-    [Unit("Per Sec")]public int ExtractionRate;
+    [Unit("Per Sec")]public float ExtractionRate;
 
 
 }
